refactor(testing): split createTestDepot.ts and testSaga.ts (#71)
